added multiple methods and some improvements to existing methods (#51)
* Update strings.py

* Adding the needed modifications for DeleteMessage command

* Update bot.py

* added multiple methods and some improvements to existing methods

send_message and send_media can now return message id if local_id was provided. This is done for cases when you might need to use the message id for editing or deleting the message, or saving it in the database. Example: z = await bot.send_message(chat.id, "Hello")
added methods send_media_by_id and edit_message_reply_markup
added inline_command_rows
changed the uiState.py so that the keyboards are not reset each time a new message is sent